day2/requirements.txt
python-dlt[bigquery] >= 0.2.0a14
